Refactor files. Clean up runtime warnings
